refactor: enhance usecase error handling

+++ internal/usecase/todo_usecase.go
@@ -1,7 +1,10 @@
 package usecase
 
 import (
+	"fmt"
+
 	"github.com/kei3dev/todo-app-api-go/internal/entity"
+	"github.com/kei3dev/todo-app-api-go/internal/errors"
 	"github.com/kei3dev/todo-app-api-go/internal/repository"
 )
 
@@ -26,7 +29,14 @@ func (u *todoUsecaseImpl) CreateTodo(todo *entity.Todo) error {
 }
 
 func (u *todoUsecaseImpl) GetTodoByID(id uint) (*entity.Todo, error) {
-	return u.todoRepo.FindByID(id)
+	todo, err := u.todoRepo.FindByID(id)
+	if err != nil {
+		return nil, fmt.Errorf("failed to find todo: %w", err)
+	}
+	if todo == nil {
+		return nil, errors.ErrTodoNotFound
+	}
+	return todo, nil
 }
 
 func (u *todoUsecaseImpl) GetTodosByUserID(userID uint) ([]entity.Todo, error) {
@@ -34,9 +44,29 @@ func (u *todoUsecaseImpl) GetTodosByUserID(userID uint) ([]entity.Todo, error) {
 }
 
 func (u *todoUsecaseImpl) UpdateTodo(todo *entity.Todo) error {
+	existingTodo, err := u.todoRepo.FindByID(todo.ID)
+	if err != nil {
+		return fmt.Errorf("failed to find todo: %w", err)
+	}
+	if existingTodo == nil {
+		return errors.ErrTodoNotFound
+	}
+
+	if existingTodo.UserID != todo.UserID {
+		return errors.ErrUnauthorized
+	}
+
 	return u.todoRepo.Update(todo)
 }
 
 func (u *todoUsecaseImpl) DeleteTodo(id uint) error {
+	todo, err := u.todoRepo.FindByID(id)
+	if err != nil {
+		return fmt.Errorf("failed to find todo: %w", err)
+	}
+	if todo == nil {
+		return errors.ErrTodoNotFound
+	}
+
 	return u.todoRepo.Delete(id)
 }
